requirements.txt
numpy>=1.21.0
matplotlib>=3.5.0
scikit-learn>=1.0.0
seaborn>=0.11.0
Pillow>=8.3.0
jupyter>=1.0.0
opencv-python>=4.5.0